metodos cambiar estado y buscar por id

+++ src/Controllers/UsersController.php
@@ -10,7 +10,7 @@ use App\Controllers\BaseController;
 class UsersController extends BaseController
 {
 
-    public function getAll($request, $response, $args)
+    public function getAll(Request $request, Response $response, $args)
     {
         $payload = $this->container->get('users_service')->getAll();
 
@@ -20,14 +20,31 @@ class UsersController extends BaseController
             ->withStatus(200);
     }
 
-    public function select($request, $response, $args)
+    public function select(Request $request, Response $response, $args)
     {
-        $payload = $this->container->get('users_service')->select($args);
+        //se verifica que venga un JSON
+        $contentType = $request->getHeaderLine('Content-Type');
+        if (strstr($contentType, 'application/json')) {
+            //$contents = json_decode(file_get_contents('php://input'), true);//Se parsea la data en array
+            $contents = json_decode(file_get_contents('php://input')); //Se parsea la data en obj
+            if (json_last_error() === JSON_ERROR_NONE) {
+                $request = $request->withParsedBody($contents);
+            }
+        }
 
+        if ($contents->id != "") {
+            //se envia el obj al modelo
+            $payload = $this->container->get('users_service')->select($contents);
+            $status = 200;
+        } else {
+            $payload = json_encode("Complete todos los campos.");
+            $status = 400;
+        }
+        //se envia la respuesta del payload
         $response->getBody()->write($payload);
         return $response
             ->withHeader('content-type', 'application/json')
-            ->withStatus(200);
+            ->withStatus($status);
     }
 
     public function insert(Request $request, Response $response, $args)
@@ -41,16 +58,23 @@ class UsersController extends BaseController
                 $request = $request->withParsedBody($contents);
             }
         }
-        //se envia el obj al modelo
-        $payload = $this->container->get('users_service')->insert($contents);
+
+        if ($contents->nombre != "") {
+            //se envia el obj al modelo
+            $payload = $this->container->get('users_service')->insert($contents);
+            $status = 200;
+        } else {
+            $payload = json_encode("Complete todos los campos.");
+            $status = 400;
+        }
         //se envia la respuesta del payload
         $response->getBody()->write($payload);
         return $response
             ->withHeader('content-type', 'application/json')
-            ->withStatus(200);
+            ->withStatus($status);
     }
 
-    public function update($request, $response, $args)
+    public function update(Request $request, Response $response, $args)
     {
         //$id = $request->getAttribute('id');//para solicitar datos de la url
         //se verifica que venga un JSON
@@ -62,16 +86,51 @@ class UsersController extends BaseController
                 $request = $request->withParsedBody($contents);
             }
         }
-        //se envia el obj al modelo
-        $payload = $this->container->get('users_service')->update($contents);
+
+        if ($contents->id != "" || $contents->nombre != "") {
+            //se envia el obj al modelo
+            $payload = $this->container->get('users_service')->update($contents);
+            $status = 200;
+        } else {
+            $payload = json_encode("Complete todos los campos.");
+            $status = 400;
+        }
         //se envia la respuesta del payload
         $response->getBody()->write($payload);
         return $response
             ->withHeader('content-type', 'application/json')
-            ->withStatus(200);
+            ->withStatus($status);
     }
 
-    public function delete($request, $response, $args)
+    public function updateState(Request $request, Response $response, $args)
+    {
+        //$id = $request->getAttribute('id');//para solicitar datos de la url
+        //se verifica que venga un JSON
+        $contentType = $request->getHeaderLine('Content-Type');
+        if (strstr($contentType, 'application/json')) {
+            //$contents = json_decode(file_get_contents('php://input'), true);//Se parsea la data en array
+            $contents = json_decode(file_get_contents('php://input')); //Se parsea la data en obj
+            if (json_last_error() === JSON_ERROR_NONE) {
+                $request = $request->withParsedBody($contents);
+            }
+        }
+
+        if ($contents->id != "" || $contents->estado != "") {
+            //se envia el obj al modelo
+            $payload = $this->container->get('users_service')->updateState($contents);
+            $status = 200;
+        } else {
+            $payload = json_encode("Complete todos los campos.");
+            $status = 400;
+        }
+        //se envia la respuesta del payload
+        $response->getBody()->write($payload);
+        return $response
+            ->withHeader('content-type', 'application/json')
+            ->withStatus($status);
+    }
+
+    public function delete(Request $request, Response $response, $args)
     {
         //se verifica que venga un JSON
         $contentType = $request->getHeaderLine('Content-Type');
@@ -82,12 +141,19 @@ class UsersController extends BaseController
                 $request = $request->withParsedBody($contents);
             }
         }
-        //se envia el obj al modelo
-        $payload = $this->container->get('users_service')->delete($contents);
+
+        if ($contents->id != "") {
+            //se envia el obj al modelo
+            $payload = $this->container->get('users_service')->delete($contents);
+            $status = 200;
+        } else {
+            $payload = json_encode("Complete todos los campos.");
+            $status = 400;
+        }
         //se envia la respuesta del payload
         $response->getBody()->write($payload);
         return $response
             ->withHeader('content-type', 'application/json')
-            ->withStatus(200);
+            ->withStatus($status);
     }
 }
